extract post-login dialog tasks into service
